+++ sentry-log/src/lib.rs
@@ -11,13 +11,10 @@
 //! - Records can be captured as traditional [logs](https://docs.sentry.io/product/explore/logs/)
 //!   Logs can be viewed and queried in the Logs explorer.
 //!
-//! By default anything above `Info` is recorded as a breadcrumb and
-//! anything above `Error` is captured as error event.
-//!
-//! To capture records as Sentry logs:
-//! 1. Enable the `logs` feature of the `sentry` crate.
-//! 2. Initialize the SDK with `enable_logs: true` in your client options.
-//! 3. Set up a custom filter (see below) to map records to logs (`LogFilter::Log`) based on criteria such as severity.
+//! By default anything at or above `Info` is recorded as a breadcrumb and
+//! anything at or above `Error` is captured as error event.
+//! Additionally, if the `sentry` crate is used with the `logs` feature flag, anything at or above `Info`
+//! is captured as a [Structured Log](https://docs.sentry.io/product/explore/logs/).
 //!
 //! # Examples
 //!